requirements.txt
kafka-python==2.0.2
pandas==2.1.1
scikit-learn==1.3.0
numpy==1.26.0
joblib==1.3.2
